% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_RegionSizeR.R
\name{launch_RegionSizeR}
\alias{launch_RegionSizeR}
\title{launch_RegionSizeR}
\usage{
launch_RegionSizeR(host = "127.0.0.1", port = NULL)
}
\arguments{
\item{host}{host link (defaults to the local machine "127.0.0.1")}

\item{port}{port number (randomly chosen unless specified as a certain number)}
}
\value{
No return value.
}
\description{
Starts the RegionSizer application in the client’s browser.
}
\examples{


if(interactive()) {
    launch_RegionSizeR()
}

}
